twi as inline, header only
